backend/pyproject.toml
[project]
name = "disco2000-backend"
version = "0.1.0"
description = "Backend FastAPI pour Disco2000"
authors = [ { name = "Christophe", email = "email@example.com" } ]
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy>=2.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "passlib[argon2]>=1.7.4",
    "pyjwt>=2.8.0",
    "httpx>=0.27.0"
]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio", "pytest-cov", "ruff", "black", "mypy"]

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100
extend-select = ["I"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
